Port: the formatter compiled for 16.7s against a 30s test budget
83 tests in the suite give `cc` 30 seconds to build the port.
pc_port/mips_formatter.c took 16.7 of them at -O2 on an idle machine, so
under any load the build blew the budget and the test reported
subprocess.TimeoutExpired. That is what the 39 failures and 11 errors in
the 2026-09-14 full run were: compile timeouts, not regressions. A suite
that is red for a reason unrelated to the code under test has stopped
being evidence, which is the actual damage here.

-ftime-report put 83% of the time in the dominator pass and
-fdisable-tree-dom1/2/3=<function> narrowed it to one function. Three
things were wrong and all three were the same mistake: machine-generated
*code* where the data belonged in a *table*.

- merge_pending_matches held six `||` chains, 225 `(pc && word && reg)`
  terms in single boolean expressions. Folded into kMergeRetireSites, 118
  unique rows. 14.2s -> 1.8s on its own.
- formatter_fetch held 2,077 `else if (pc >= a && pc < b)` branches across
  17 runs. Each run of consecutive branches with one guard and disjoint
  ranges became a sorted table and a binary search. That function went
  from 5,659 lines to 579.
- Eight GTE site scans each carried their own copy of the loop; with the
  element count a visible constant GCC unrolls it and the dominator pass
  explores the unrolled paths. They now share one noinline lookup, which
  keeps the trip count opaque.

16.7s -> 1.43s, and the file went from 27,087 to 22,356 lines.
tests/test_pvd_irq.py went from 4 failures to 5 passes.

This was never only about the clock. The dispatch was O(n) in carved
ranges at run time too, and the seam has to reach all 3,962 functions,
about three times today's count.

Both rewrites are proven equivalent, not reviewed and hoped over. Old and
new translation units were built as two shared objects and compared
directly:

- formatter_fetch: every 4-byte-aligned address in [80000000,80200000)
  under all 16 combinations of the four overlay selections. 8,388,608
  fetches, 0 differences.
- merge_pending_matches: every (pc, word, reg) triple named anywhere in the
  old file, crossed with every merge kind, register, pending, delay-slot
  and branch state, plus deliberately mismatched words. 162,965 cases, 0
  differences.

A binary search only equals an ordered chain when the ranges are disjoint,
so the transform refuses any run containing an overlap. The SC01/0012
overlay alias at 80128158 is still ordered branches, as is every branch
carrying a condition beyond its pc range.

Also registers 2 drafts the repair search matched. Coverage 248,016 ->
248,384 / 806,272 (30.76% -> 30.81%), verify_registry green at 2,649/2,649.
The native menu is still NOT_REACHED.

+++ src/overlays/main_0012/8017b0e4.c
@@ -82,11 +82,11 @@ extern s16 D_801853BC;
 extern s16 D_801853C4;
 
 void func_8017B0E4(void *arg0, s16 arg1) {
-    M2C_FIELD(arg0, s16 *, 0x48) = (s32) func_80012C6C(M2C_FIELD(arg0, s16 *, 0x48), D_801853C4, arg1);
-    M2C_FIELD(arg0, s16 *, 0x4C) = (s32) func_80012C6C(M2C_FIELD(arg0, s16 *, 0x4C), D_801853C6, arg1);
-    M2C_FIELD(arg0, s16 *, 0x50) = (s32) func_80012C6C(M2C_FIELD(arg0, s16 *, 0x50), D_801853C8, arg1);
-    M2C_FIELD(arg0, s16 *, 0x3C) = (s32) func_80012C6C(M2C_FIELD(arg0, s16 *, 0x3C), D_801853BC, arg1);
-    M2C_FIELD(arg0, s16 *, 0x40) = (s32) func_80012C6C(M2C_FIELD(arg0, s16 *, 0x40), D_801853BE, arg1);
-    M2C_FIELD(arg0, s16 *, 0x44) = (s32) func_80012C6C(M2C_FIELD(arg0, s16 *, 0x44), D_801853C0, arg1);
+    M2C_FIELD(arg0, s32 *, 0x48) = (s32) func_80012C6C(M2C_FIELD(arg0, s32 *, 0x48), D_801853C4, arg1);
+    M2C_FIELD(arg0, s32 *, 0x4C) = (s32) func_80012C6C(M2C_FIELD(arg0, s32 *, 0x4C), D_801853C6, arg1);
+    M2C_FIELD(arg0, s32 *, 0x50) = (s32) func_80012C6C(M2C_FIELD(arg0, s32 *, 0x50), D_801853C8, arg1);
+    M2C_FIELD(arg0, s32 *, 0x3C) = (s32) func_80012C6C(M2C_FIELD(arg0, s32 *, 0x3C), D_801853BC, arg1);
+    M2C_FIELD(arg0, s32 *, 0x40) = (s32) func_80012C6C(M2C_FIELD(arg0, s32 *, 0x40), D_801853BE, arg1);
+    M2C_FIELD(arg0, s32 *, 0x44) = (s32) func_80012C6C(M2C_FIELD(arg0, s32 *, 0x44), D_801853C0, arg1);
 }
 #endif
